requirements.txt
ipykernel
ipywidgets
IProgress

python-dotenv
transformers
torch
tqdm
wandb

pytest